[Feat] #114 뷰모드 진입 시  스크롤 구현
선택한 날짜의 선택한 페이지로 스크롤합니다.
